Queue introduced

+++ CMD/cleanup.sh
@@ -1,4 +1,8 @@
 cd `readlink -f $0 | xargs dirname`
-cd ..
-sudo rm -R simjob/* && echo Cleaned robo logs
-sudo rm -R robo/container/* && echo Cleaned sagemaker temp
+if [ "$1" = "nosudo" ]; then
+	./nosudo-clean
+else
+	cd ..
+	sudo rm -R robo/job/* && echo Cleaned robo logs
+	sudo rm -R robo/container/* && echo Cleaned sagemaker temp
+fi
